copy to clipboard fix applied again
