+++ app/labs/page.tsx
@@ -11,6 +11,7 @@ import { filterLabs, getPersonas, getTopics, getAreas, getProblems } from "@/lib
 import SelectFilter from "@/components/labs/SelectFilter";
 import { getPersonaIntro } from "@/data";
 import { downloadLabSheet } from "@/lib/download";
+import { copyToClipboard } from "@/lib/utils";
 import CaseStudyHierarchy from "@/components/labs/CaseStudyHierarchy";
 
 const findMatchingOption = (options: string[], urlValue: string): string => {
@@ -174,27 +175,6 @@ const LabsPage = () => {
   const handleDownload = () => {
     if (!selectedLab) return;
     downloadLabSheet(selectedLab, personaIntro, selectedHierarchicalData, topicHierarchy);
-  };
-
-  const copyToClipboard = (text: string) => {
-    if (navigator.clipboard) {
-      navigator.clipboard.writeText(text).catch(() => {
-        fallbackCopy(text);
-      });
-    } else {
-      fallbackCopy(text);
-    }
-  };
-
-  const fallbackCopy = (text: string) => {
-    const textarea = document.createElement("textarea");
-    textarea.value = text;
-    textarea.style.position = "fixed";
-    textarea.style.opacity = "0";
-    document.body.appendChild(textarea);
-    textarea.select();
-    document.execCommand("copy");
-    document.body.removeChild(textarea);
   };
 
   return (
